+++ packages/network-controller/tests/provider-api-tests/block-param.ts
@@ -1,4 +1,4 @@
-import { rpcErrors } from '@metamask/rpc-errors';
+import { errorCodes, rpcErrors } from '@metamask/rpc-errors';
 import type { Hex } from '@metamask/utils';
 
 import type { MockRequest, ProviderType } from './helpers';
@@ -10,6 +10,7 @@ import {
   withNetworkClient,
 } from './helpers';
 import { testsForRpcFailoverBehavior } from './rpc-failover';
+import { CUSTOM_RPC_ERRORS } from '../../src/rpc-service/rpc-service';
 import { NetworkClientType } from '../../src/types';
 
 type TestsForRpcMethodSupportingBlockParam = {
@@ -429,12 +430,19 @@ export function testsForRpcMethodSupportingBlockParam(
     });
 
     describe.each([
-      [405, 'The method does not exist / is not available.'],
-      [429, 'Request is being rate limited.'],
+      [401, CUSTOM_RPC_ERRORS.unauthorized],
+      [402, errorCodes.rpc.resourceUnavailable],
+      [404, errorCodes.rpc.resourceUnavailable],
+      [422, CUSTOM_RPC_ERRORS.httpClientError],
+      [429, errorCodes.rpc.limitExceeded],
     ])(
       'if the RPC endpoint returns a %d response',
-      (httpStatus, errorMessage) => {
-        it('throws a custom error', async () => {
+      (httpStatus, rpcErrorCode) => {
+        const expectedError = expect.objectContaining({
+          code: rpcErrorCode,
+        });
+
+        it('throws a custom error without retrying the request', async () => {
           await withMockedCommunications({ providerType }, async (comms) => {
             const request = {
               method,
@@ -466,84 +474,90 @@ export function testsForRpcMethodSupportingBlockParam(
               async ({ makeRpcCall }) => makeRpcCall(request),
             );
 
-            await expect(promiseForResult).rejects.toThrow(errorMessage);
+            await expect(promiseForResult).rejects.toThrow(expectedError);
           });
+        });
+
+        // NOTE: We do not test the RPC failover behavior here because only 5xx
+        // errors break the circuit and cause a failover.
+      },
+    );
+
+    describe.each([500, 501, 505, 506, 507, 508, 510, 511])(
+      'if the RPC endpoint returns a %d response',
+      (httpStatus) => {
+        const expectedError = expect.objectContaining({
+          code: errorCodes.rpc.resourceUnavailable,
+        });
+
+        it('throws a generic, undescriptive error', async () => {
+          await withMockedCommunications({ providerType }, async (comms) => {
+            const request = {
+              method,
+              params: buildMockParams({ blockParam, blockParamIndex }),
+            };
+
+            // The first time a block-cacheable request is made, the
+            // block-cache middleware will request the latest block number
+            // through the block tracker to determine the cache key. Later,
+            // the block-ref middleware will request the latest block number
+            // again to resolve the value of "latest", but the block number is
+            // cached once made, so we only need to mock the request once.
+            comms.mockNextBlockTrackerRequest({ blockNumber: '0x100' });
+            // The block-ref middleware will make the request as specified
+            // except that the block param is replaced with the latest block
+            // number.
+            comms.mockRpcCall({
+              request: buildRequestWithReplacedBlockParam(
+                request,
+                blockParamIndex,
+                '0x100',
+              ),
+              response: {
+                httpStatus,
+              },
+            });
+            const promiseForResult = withNetworkClient(
+              { providerType },
+              async ({ makeRpcCall }) => makeRpcCall(request),
+            );
+
+            await expect(promiseForResult).rejects.toThrow(expectedError);
+          });
+        });
+
+        testsForRpcFailoverBehavior({
+          providerType,
+          requestToCall: {
+            method,
+            params: buildMockParams({ blockParam, blockParamIndex }),
+          },
+          getRequestToMock: (request: MockRequest, blockNumber: Hex) => {
+            return buildRequestWithReplacedBlockParam(
+              request,
+              blockParamIndex,
+              blockNumber,
+            );
+          },
+          failure: {
+            httpStatus,
+          },
+          isRetriableFailure: false,
+          getExpectedError: () => expectedError,
+          getExpectedBreakError: () =>
+            expect.objectContaining({
+              message: `Fetch failed with status '${httpStatus}'`,
+            }),
         });
       },
     );
 
-    describe('if the RPC endpoint returns a response that is not 405, 429, 503, or 504', () => {
-      const httpStatus = 500;
-      const errorMessage = `Non-200 status code: '${httpStatus}'`;
-
-      it('throws a generic, undescriptive error', async () => {
-        await withMockedCommunications({ providerType }, async (comms) => {
-          const request = {
-            method,
-            params: buildMockParams({ blockParam, blockParamIndex }),
-          };
-
-          // The first time a block-cacheable request is made, the
-          // block-cache middleware will request the latest block number
-          // through the block tracker to determine the cache key. Later,
-          // the block-ref middleware will request the latest block number
-          // again to resolve the value of "latest", but the block number is
-          // cached once made, so we only need to mock the request once.
-          comms.mockNextBlockTrackerRequest({ blockNumber: '0x100' });
-          // The block-ref middleware will make the request as specified
-          // except that the block param is replaced with the latest block
-          // number.
-          comms.mockRpcCall({
-            request: buildRequestWithReplacedBlockParam(
-              request,
-              blockParamIndex,
-              '0x100',
-            ),
-            response: {
-              httpStatus,
-            },
-          });
-          const promiseForResult = withNetworkClient(
-            { providerType },
-            async ({ makeRpcCall }) => makeRpcCall(request),
-          );
-
-          await expect(promiseForResult).rejects.toThrow(errorMessage);
-        });
-      });
-
-      testsForRpcFailoverBehavior({
-        providerType,
-        requestToCall: {
-          method,
-          params: buildMockParams({ blockParam, blockParamIndex }),
-        },
-        getRequestToMock: (request: MockRequest, blockNumber: Hex) => {
-          return buildRequestWithReplacedBlockParam(
-            request,
-            blockParamIndex,
-            blockNumber,
-          );
-        },
-        failure: {
-          httpStatus,
-        },
-        isRetriableFailure: false,
-        getExpectedError: () =>
-          expect.objectContaining({
-            message: errorMessage,
-          }),
-        getExpectedBreakError: () =>
-          expect.objectContaining({
-            message: `Fetch failed with status '500'`,
-          }),
-      });
-    });
-
-    describe.each([503, 504])(
+    describe.each([502, 503, 504])(
       'if the RPC endpoint returns a %d response',
       (httpStatus) => {
-        const errorMessage = 'Gateway timeout';
+        const expectedError = expect.objectContaining({
+          code: errorCodes.rpc.resourceUnavailable,
+        });
 
         it('retries the request up to 5 times until there is a 200 response', async () => {
           await withMockedCommunications({ providerType }, async (comms) => {
@@ -640,7 +654,7 @@ export function testsForRpcMethodSupportingBlockParam(
                 );
               },
             );
-            await expect(promiseForResult).rejects.toThrow(errorMessage);
+            await expect(promiseForResult).rejects.toThrow(expectedError);
           });
         });
 
@@ -661,10 +675,7 @@ export function testsForRpcMethodSupportingBlockParam(
             httpStatus,
           },
           isRetriableFailure: true,
-          getExpectedError: () =>
-            expect.objectContaining({
-              message: expect.stringContaining(`Gateway timeout`),
-            }),
+          getExpectedError: () => expectedError,
           getExpectedBreakError: () =>
             expect.objectContaining({
               message: expect.stringContaining(
@@ -803,7 +814,9 @@ export function testsForRpcMethodSupportingBlockParam(
     );
 
     describe('if the RPC endpoint responds with invalid JSON', () => {
-      const errorMessage = 'not valid JSON';
+      const expectedError = expect.objectContaining({
+        code: errorCodes.rpc.parse,
+      });
 
       it('retries the request up to 5 times until it responds with valid JSON', async () => {
         await withMockedCommunications({ providerType }, async (comms) => {
@@ -900,7 +913,7 @@ export function testsForRpcMethodSupportingBlockParam(
             },
           );
 
-          await expect(promiseForResult).rejects.toThrow(errorMessage);
+          await expect(promiseForResult).rejects.toThrow(expectedError);
         });
       });
 
@@ -921,9 +934,10 @@ export function testsForRpcMethodSupportingBlockParam(
           body: 'invalid JSON',
         },
         isRetriableFailure: true,
-        getExpectedError: () =>
+        getExpectedError: () => expectedError,
+        getExpectedBreakError: () =>
           expect.objectContaining({
-            message: expect.stringContaining(errorMessage),
+            message: expect.stringContaining('invalid json'),
           }),
       });
     });